Implement itertive method for cluster node relaxing.
This implementation is more explicit in terms of how it is moving the cluster
node towards destination. In addition, if the destination is too far from the
current location, we bring it back to at most 2*cluster_radius. This can be
done, because any larger movement of the cluster node will break the
constraints.