Chose a closer destination when relaxing.
When chosing a destination point, limit the point to the line segment
connecting previous node to the next node. In some cases, the closest
point on the line passing through these points is outside the segment,
and moving cluster center towards this point is not the best we can do.